Update autonote.py
removed the lines to add the topic sentence because sometimes the first sentence is not that important
create an issue if you feel like i should change it back
also if you have a way to split sentences that would be great. maybe i'll figure that out in the next commit. (start a pull request)