removed nltk dependency
Earlier, I used `nltk` to get the list of stopwords and the sentence and word tokenizers. However, I've figured out how to implement the sentence and word tokenizers with only the Python standard libary.

I also found the list of stopwords and stored it in `STOPWORDS`.

added gitignore for config files