changed paragraph rules
i made it so that paragraphs are now separated by two `\n`s instead of one. This is to allow text to be copied from the book, where line breaks are common because of the text's formatting. You might still have to add line breaks though...
I'll work on a system to check for indents for paragraphs.

Meanwhile, if anyone has a better idea of how to split sentences, I'd love to hear it, because the regex-based solution reorders the text for some reason. (Try to use only the standard library, just for easy porting or whatever.)